resolve some more of the warnings when running tests

+++ woodwork/tests/fixtures/datetime_freq.py
@@ -50,8 +50,8 @@ TAIL_RANGE_LEN = 100
 def pad_datetime_series(dates, freq, pad_start=0, pad_end=100):
     dates = [pd.Timestamp(d) for d in dates]
 
-    head = pd.Series([])
-    tail = pd.Series([])
+    head = pd.Series([], dtype="float64")
+    tail = pd.Series([], dtype="float64")
 
     if pad_start > 0:
         head = (
